refactor(exceptions): split the module into a package by error family
Each family gets its own submodule over a shared `base` and `rendering`;
`__init__` re-exports every public name, so `from modern_di import exceptions`
and `from modern_di.exceptions import X` are unchanged.

The class bodies are a verbatim move: all 37 top-level blocks keep byte-identical
ASTs and the 34 public names are the same set. `__module__` now names the family
submodule, so it reads that way in tracebacks.

Closes #473
